Updated pages for gallery
I have updated the pages for the gallery and have created new pages that when clicked on load up a new grid using javascript. I've also changed the header class to ids for simplicity as I may use the active class in the near future, causing issues with each element.