I've added javascript and a date function
I decided to add a simle function to determine the date and an if else condition to tell the user if they'll have volleyball today. I also increased how specific the classes/ids are for the li elements and have altered the button hover effect to appear both more unique and responsive/interactive.