Page gallery numbers fix
There was a bug with the page numbers where it wouldn't allow users to view older pages after page 4 due to their being no link.

The solution to this problem in the easiest light would be to provide more viewing options for users, I've provided initially, up to 10 of the links, after page 10, it inlcudes pages 9-19.